Ghost fetching, gps accuracy and home screen ux
Previously:

* Ghosts were being cacahed too long.
* In door gps accuracy was causing bouncing of points
* Home screen, upload dial had a different ux

_The problem with GPS accuracy:_ the gps stream returns a reading with an
embedded error range. Sometimes this error range is high, eg 20-100m. We
were previously ignoring this data. Indoors, it is prudent to only take
notice of readings that have a low error range.

This does mean that if the error range is always high, we will need to
make a decision. This will manifest as points simply not showing up on
screen. We need to look out for this in the field.

A better way to handle this would be some sort of moving average over
the points. If the error is consistently high, but the user is moving,
we can take an average of the window.

_The problem with ghost caching:_ the ghosts embedded in sites.json shold
be refreshed based on the filename. In the old code, there was a bug
where we would download the ghost for each site under a predetermined
name and use that name to check whether we should re-fetch it.

Now, we save the file locally under the same site/remote_file_name and
use this remote_file_name change to determine whether we should re-get
it. If the remote_file_name remains the same, we don't bother
downloading it (i.e we compare the name in sites.json with the
downloaded file).