Fix GPS acquisition / upload race
There was a race condition where we couldn't detect any nearby sites,
and hence we failed to set the bucket root on the new sites json. The
symptom of this bug is that the site will how up in the add site flow
without any questions. While this is OK in the sense that we will not
always have a "nearby site" from which to copy info (even when there is
a gps, because this might be the very first site being added on the app)
- we should have a reliable fallback option. We should never be in a
  situation where there is a site, but no home for that site's data in
the backend.

This PR fixes the issue by having a resolution of fallback options for
malformed sites:
1. Look at the first site's bucket root
2. If there is no other site, user the default fomogues bucket

Also bumps the pubspec.yaml version.