Fix GPS acquisition / upload race
There was a race condition where we couldn't detect any nearby sites,
and hence we failed to set the bucket root on the new sites json. The
symptom of this bug is that the site will how up in the add site flow
without any questions. While this is OK in the sense that we will not
always have a "nearby site" from which to copy info (even when there is
a gps, because this might be the very first site being added on the app)
- we should have a reliable fallback option. We should never be in a
situation where there is a site, but no home for that site's data in
the backend.

This PR fixes the issue by disallowing the button press when the user
position and nearest site haven't been computed. When these two have
been computed but the user is not within radius of the nearest site, the
button still leads to new site creation.

This commit also changes the trigger radius from 50 to 30m based on
field testing.